# area-node/OpenAPI.yaml
openapi: 3.0.3
info:
  title: AREA API - OpenAPI 3.0
  description: |-
    This is the AREA Server API description based on the OpenAPI 3.0 specification.  You can find out more about
    AREA at [GitHub repository](https://github.com/christophechr/LeAREA/tree/main)

  contact:
    email: raphael.mercie@epitech.eu
  version: 1.0.11
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /triggers:
    get:
      tags:
        - Triggers
      summary: Get all possible triggers
      description: get all available triggers for each available services, and what parameter you need to create them
      operationId: triggers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: |
            Returns the available triggers as JSON object.<br />
            - <b>id</b> The ID of the service
            - <b>name</b> The name of the service to display (e.g. GitHub)
            - <b>triggers</b> List of all triggers available for the service
              - <b>id</b> ID of the trigger
              - <b>name</b> Name of the trigger to display (e.g. When temperature is...)
              - <b>description</b> A brief to describe the trigger check
              - <b>params</b> Parameters accepted by the trigger
                - <b>id</b> ID of the trigger
                - <b>name</b> Parameter displayable name
                - <b>type</b> Type of the parameter. It can be : <b>string</b>, <b>number</b>, <b>enum</b> or <b>boolean</b>
                - <b>required</b> If the parameter is required or not
                - <b>description</b> A brief to describe the parameter
                - <b>values</b> OPTIONAL : If the parameter is an enum, the list of possible values
                  - <b>name</b> Name of the value to display
                  - <b>value</b> Value of the enum value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggersConfig'
        '401':
          description: Unauthorized
  /actions:
    get:
      tags:
        - Actions
      summary: Get all possible actions
      description: get all available actions for each available services, and what parameter you need to create them
      operationId: actions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: |
            Returns the available actions as JSON object.<br />
            - <b>id</b> The ID of the service
            - <b>name</b> The name of the service to display (e.g. GitHub)
            - <b>actions</b> List of all actions available for the service
              - <b>id</b> ID of the action
              - <b>name</b> Name of the action to display (e.g. When temperature is...)
              - <b>description</b> A brief to describe the action check
              - <b>params</b> Parameters accepted by the action
                - <b>id</b> ID of the action
                - <b>name</b> Parameter displayable name
                - <b>type</b> Type of the parameter. It can be : <b>string</b>, <b>number</b>, <b>enum</b> or <b>boolean</b>
                - <b>required</b> If the parameter is required or not
                - <b>description</b> A brief to describe the parameter
                - <b>values</b> OPTIONAL : If the parameter is an enum, the list of possible values
                  - <b>name</b> Name of the value to display
                  - <b>value</b> Value of the enum value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsConfig'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
      bearerAuth:            # arbitrary name for the security scheme
        type: http
        scheme: bearer
        bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  schemas:
...
    TriggersConfig:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
            description: The ID of the trigger service.
          name:
            type: string
            description: The name of the trigger service.
          actions:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                  description: The ID of the trigger.
                name:
                  type: string
                  description: The name of the trigger.
                description:
                  type: string
                  description: Description of the trigger.
                params:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        description: The ID of the parameter.
                      name:
                        type: string
                        description: The name of the parameter.
                      type:
                        type: string
                        description: The data type of the parameter.
                      required:
                        type: boolean
                        description: Indicates if the parameter is required.
                      description:
                        type: string
                        description: Description of the parameter.
    ActionsConfig:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
            description: The ID of the action service.
          name:
            type: string
            description: The name of the action service.
          actions:
            type: array
            items:
              type: object
              properties:
                id:
                  type: string
                  description: The ID of the action action.
                name:
                  type: string
                  description: The name of the action action.
                description:
                  type: string
                  description: Description of the action action.
                params:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        description: The ID of the parameter.
                      name:
                        type: string
                        description: The name of the parameter.
                      type:
                        type: string
                        description: The data type of the parameter.
                      required:
                        type: boolean
                        description: Indicates if the parameter is required.
                      description:
                        type: string
                        description: Description of the parameter.